refactor(phase6): extract list-query helpers to codegen/list_tokens.rs
Moves the 462-line list-query helper block (ListBindKind, filter
condition builders, cursor/plan/response ident helpers, bind-match
token generators) from codegen.rs into its own submodule.

codegen.rs: 1,717 → 1,255 lines (−27 %).  All 233 tests pass.
